+++ src/main/java/com/analitrix/sellbook/entity/Book.java
@@ -2,15 +2,8 @@ package com.analitrix.sellbook.entity;
 
 
 import java.util.Date;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
+
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -33,7 +26,7 @@ public class Book {
 	private boolean available;
 	private Long units;
 	private String editorial;
-	private double cost;
+	private Long cost;
 	private String author;
 	private String image;
 	@Temporal(TemporalType.TIMESTAMP)
